move fourier helper to plotting utils
